+++ src/Former/Former.php
@@ -39,6 +39,11 @@ class Former
    */
   private static $rules = array();
 
+  /**
+   * The namespace of fields
+   */
+  const FIELDSPACE = 'Former\Fields\\';
+
   ////////////////////////////////////////////////////////////////////
   //////////////////////////// INTERFACE /////////////////////////////
   ////////////////////////////////////////////////////////////////////
@@ -55,8 +60,14 @@ class Former
     // Form opener
     if (str_contains($method, 'open')) {
       static::$form = new Form;
+      static::form()->open($method, $parameters);
 
-      return static::form()->open($method, $parameters);
+      return new static;
+    }
+
+    // Avoid conflict with chained label method
+    if ($method == 'label') {
+      return call_user_func_array('static::_label', $parameters);
     }
 
     // Checking for any supplementary classes
@@ -67,10 +78,14 @@ class Former
     static::$field = null;
 
     // Picking the right class
-    if (class_exists('\Former\Fields\\'.ucfirst($method))) {
+    if (class_exists(static::FIELDSPACE.ucfirst($method))) {
       $callClass = ucfirst($method);
     } else {
       switch ($method) {
+        case 'submit':
+        case 'reset':
+          $callClass = 'Button';
+          break;
         case 'multiselect':
           $callClass = 'Select';
           break;
@@ -89,8 +104,13 @@ class Former
       }
     }
 
+    // Check for potential errors
+    if (!class_exists(static::FIELDSPACE.$callClass)) {
+      throw new \Exception('The class "' .static::FIELDSPACE.$callClass. '" called by field "' .$method. '" doesn\'t exist');
+    }
+
     // Listing parameters
-    $class = '\Former\Fields\\'.$callClass;
+    $class = static::FIELDSPACE.$callClass;
     static::$field = new $class(
       $method,
       array_get($parameters, 0),
@@ -107,11 +127,18 @@ class Former
       static::$field->inline();
     }
 
-    // Add any size we found
-    $size = Framework::getFieldSizes($classes);
-    if($size) static::$field->addClass($size);
+    // Filter classes according to field type
+    $classes = $callClass == 'Button'
+      ? Framework::getButtonTypes($classes)
+      : Framework::getFieldSizes($classes);
 
-    return new Former;
+    // Add any supplementary classes we found
+    if($classes) static::$field->addClass($classes);
+
+    // As Buttons are more of a helper class, we return them directly
+    if($callClass == 'Button') return static::$field;
+
+    return new static;
   }
 
   /**
@@ -123,9 +150,13 @@ class Former
    */
   public function __call($method, $parameters)
   {
-    $object = method_exists($this->control(), $method)
-      ? $this->control()
-      : static::$field;
+    if (!static::form()->isOpened() and static::$form) {
+      $object = static::$form;
+    } else {
+      $object = method_exists($this->control(), $method)
+        ? $this->control()
+        : static::$field;
+    }
 
     // Call the function on the corresponding class
     call_user_func_array(array($object, $method), $parameters);
@@ -151,9 +182,12 @@ class Former
    */
   public function __toString()
   {
+    if (static::$form and !static::$form->isOpened()) {
+      return static::form()->__toString();
+    }
+
     // Dry syntax (hidden fields, plain fields)
-    if (static::$field->type == 'hidden' or
-        static::form()->type == 'search' or
+    if (static::$field->isUnwrappable() or
         static::form()->type == 'inline') {
           $html = static::$field->__toString();
     }
@@ -213,7 +247,7 @@ class Former
   public static function getValue($name, $fallback = null)
   {
     // Object values
-    if(is_object(static::$values)) {
+    if (is_object(static::$values)) {
 
       // Transform the name into an array
       $value = static::$values;
@@ -228,8 +262,10 @@ class Former
         }
       }
 
+      // $name  = str_contains($name, '.') ? explode('.', $name) : (array) $name;
+
       // Dive into the model
-      foreach($name as $k => $r) {
+      foreach ($name as $k => $r) {
 
         $indexable = false;
 
@@ -252,9 +288,7 @@ class Former
         }
 
         // Single model relation
-        if(isset($value->$r) || isset($value->{'get_'.$r})) {          
-          $value = $value->$r;
-        } 
+        if(isset($value->$r) or method_exists($value, 'get_'.$r)) $value = $value->$r;
         else {
           $value = $fallback;
           break;
@@ -406,7 +440,7 @@ class Former
     $buttons = func_get_args();
 
     $actions  = '<div class="form-actions">';
-    $actions .= is_array($buttons) ? implode(' ', $buttons) : $buttons;
+      $actions .= implode(' ', (array) $buttons);
     $actions .= '</div>';
 
     return $actions;
@@ -424,7 +458,9 @@ class Former
    */
   public static function getErrors($name = null)
   {
+    // Get name and translate array notation
     if(!$name) $name = static::$field->name;
+    $name = preg_replace('/\[([a-z]+)\]/', '.$1', $name);
 
     if (static::$errors) {
       return static::$errors->first($name);