+++ src/Former/Framework.php
@@ -26,6 +26,16 @@ class Framework
   );
 
   /**
+   * The button types available
+   * @var array
+   */
+  private static $types = array(
+    'bootstrap' => array(
+      'large', 'small', 'mini',
+      'block', 'info', 'inverse', 'link', 'primary', 'success', 'warning'),
+  );
+
+  /**
    * The field sizes available
    * @var array
    */
@@ -54,7 +64,7 @@ class Framework
   public static function inlineHelp($value, $attributes = array())
   {
     // Return the correct syntax according to framework
-    switch(static::current()) {
+    switch (static::current()) {
       case 'bootstrap':
         $attributes = Helpers::addClass($attributes, 'help-inline');
         $help = '<span'.HTML::attributes($attributes).'>'.$value.'</span>';
@@ -78,7 +88,7 @@ class Framework
   public static function blockHelp($value, $attributes = array())
   {
     // Block helps are only available in Bootstrap
-    if(self::isnt('bootstrap')) return static::inlineHelp($value, $attributes);
+    if(static::isnt('bootstrap')) return static::inlineHelp($value, $attributes);
 
     $attributes = Helpers::addClass($attributes, 'help-block');
 
@@ -89,10 +99,39 @@ class Framework
   /////////////////////////// FIELD WRAPPERS /////////////////////////
   ////////////////////////////////////////////////////////////////////
 
+  /**
+   * Creates a basic icon
+   *
+   * @param string $icon       The icon id
+   * @param array  $attributes Facultative attributes
+   *
+   * @return string An icon tag
+   */
+  public static function icon($icon, $attributes = array())
+  {
+    if (static::is('bootstrap')) {
+
+      // White icon
+      if (str_contains($icon, 'white')) {
+        $icon = trim(str_replace('white', null, $icon), '-');
+        $attributes = Helpers::addClass($attributes, 'icon-white');
+      }
+
+      // Check for empty icons
+      if (!$icon) return false;
+      $attributes = Helpers::addClass($attributes, 'icon-'.$icon);
+
+      // Create icon
+      $icon = '<i'.HTML::attributes($attributes).'></i>';
+    }
+
+    return $icon;
+  }
+
   public static function getFieldClasses($field)
   {
     // Wrap field in .controls if necessary
-    if(self::is('bootstrap')) {
+    if (static::is('bootstrap')) {
       $field = '<div class="controls">' .$field. '</div>';
     }
 
@@ -107,7 +146,7 @@ class Framework
    */
   public static function getLabelClasses($attributes)
   {
-    if(static::is('bootstrap')) {
+    if (static::is('bootstrap')) {
       $attributes = Helpers::addClass($attributes, 'control-label');
     }
 
@@ -122,7 +161,7 @@ class Framework
    */
   public static function getGroupClasses($attributes)
   {
-    if(static::is('bootstrap')) {
+    if (static::is('bootstrap')) {
       $attributes = Helpers::addClass($attributes, 'control-group');
     }
 
@@ -137,23 +176,42 @@ class Framework
    */
   public static function getFieldSizes($sizes)
   {
-    if(self::is(null)) return null;
-
-    // List all available sizes
-    $available = array_get(static::$sizes, static::current(), array());
+    if(static::is(null)) return null;
 
     // Filter sizes
-    $sizes = array_intersect($available, $sizes);
+    $sizes = static::getAvailable($sizes, 'sizes');
 
     // Get size from array and format it
     $size = array_pop($sizes);
-    if($size) {
-      if(self::is('bootstrap')) $size = starts_with($size, 'span') ? $size : 'input-'.$size;
-      elseif(self::is('zurb')) $size;
+    if ($size) {
+      if(static::is('bootstrap')) $size = starts_with($size, 'span') ? $size : 'input-'.$size;
+      elseif(static::is('zurb')) $size;
       else $size = null;
     }
 
     return $size;
+  }
+
+  /**
+   * Filter a button type according to the framework
+   *
+   * @param  array $types An array of types
+   * @return array        A filtered array
+   */
+  public static function getButtonTypes($types)
+  {
+    if(static::is(null)) return null;
+
+    // Filter types
+    $types = static::getAvailable($types, 'types');
+
+    // Format classes
+    if (static::is('bootstrap')) {
+      $types = static::prependClasses($types, 'btn-');
+      $types[] = 'btn';
+    } else $types = null;
+
+    return $types;
   }
 
   /**
@@ -217,6 +275,43 @@ class Framework
   public static function isnt($framework)
   {
     return static::current() != $framework;
+  }
+
+  ////////////////////////////////////////////////////////////////////
+  //////////////////////////// HELPERS ///////////////////////////////
+  ////////////////////////////////////////////////////////////////////
+
+  /**
+   * Prepend an array of classes with a string
+   *
+   * @param  array  $classes An array of classes
+   * @param  string $class   The string to prepend them with
+   * @return array           An array of prepended classes
+   */
+  private static function prependClasses($classes, $class)
+  {
+    // Add prefix to each class
+    foreach ($classes as $key => $value) {
+      if($value != $class) $classes[$key] = $class.$value;
+    }
+
+    return $classes;
+  }
+
+  /**
+   * Get all available classes from an array of classes
+   *
+   * @param  array  $classes An array of classes to filter
+   * @param  string $from    The kind of classes to get
+   * @return array           Filtered array
+   */
+  private static function getAvailable($classes, $from)
+  {
+    // List all available classes
+    $available = array_get(static::$$from, static::current(), array());
+
+    // Filter classes
+    return array_intersect($available, $classes);
   }
 
 }